Create Assignment 1 Answers.
Ans1.  Because the computer is made  of hardware RAM, ROM, HARDISK and many devices in it and all are hardware and it is easier to represent any hardware in two statement. 

Ans2.  INTEGRATED DEVELOPMENT ENVIRONMENT. 

Ans3.  Text editor is a normal editor. And code editor also as text editor but  we write a program then it is show elements by many colors. Which makes it easier to read.And it gives many facility to programmer. 

Ans4.  First programmer write a source code of many instructions. After Run it. Run option will convert the program you wrote into a source compile link and an executable file and run it.

Ans5.A  C17-2018. 
B. Dennis Ritchi. 
C. Application software is for users. And system software is for machine which are in computer. which run the system on its own.
D. By modulas method